requirements.txt
# --- Infrastructure gRPC ---
grpcio
grpcio-tools
protobuf

# --- Specialist: PDF ---
pymupdf

# --- Specialist: Audio & Video ---
faster-whisper

# --- Specialist: Image ---
Pillow

# --- Utilities ---
psutil